requirements.txt
fastapi==0.115.11
streamlit==1.43.1
uvicorn==0.34.0
langchain==0.3.20
langchain-openai==0.3.7
langchain-community==0.3.19
openai==1.65.4
pymysql==1.1.1
sqlalchemy==2.0.38
milvus==2.3.9
pymilvus==2.5.4
redis==5.2.1
celery==5.5.2
pandas==2.2.3
numpy==1.26.4
plotly==6.0.0
altair==4.2.2
pydantic==2.10.6
cryptography==42.0.8
pyjwt==2.10.1
passlib==1.7.4
python-dotenv==1.0.1
requests==2.32.3
aiofiles==24.1.0
python-multipart==0.0.20
structlog==25.4.0
prometheus-client==0.21.1
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
black==25.1.0
flake8==7.3.0
isort==6.0.1
pyyaml==6.0.2
jinja2==3.1.6
click==8.1.8
openpyxl==3.1.5
mcp==1.9.0
